refactor: rename Agents to Team, merge org chart into Team page
- Add ## UX section to AGENTS.md requiring mobile-first design across three breakpoints
- Rename sidebar nav "Agents" → "Team" and remove separate "Org chart" link
- Replace agent card grid on Team page with org chart visualization
- Make org chart agent nodes clickable (link to agent detail page)
- Delete standalone org-chart.tsx route
- Update e2e tests for renamed nav and merged page
