+++ tests/test-run-order.json
@@ -1,46 +1,46 @@
 {
-	"packages/server/src/test/__tests__/containers.test.ts": 1516,
-	"packages/server/src/test/__tests__/job-manager.test.ts": 2565,
-	"packages/server/src/test/__tests__/mcp.test.ts": 3159,
-	"packages/server/src/test/__tests__/preferences.test.ts": 3160,
-	"packages/server/src/test/__tests__/kb-docs.test.ts": 3223,
-	"packages/server/src/test/__tests__/agents.test.ts": 3270,
-	"packages/server/src/test/__tests__/live-chat.test.ts": 3329,
-	"packages/server/src/test/__tests__/container-sync.test.ts": 3404,
-	"packages/server/src/test/__tests__/comments.test.ts": 3439,
-	"packages/server/src/test/__tests__/skills.test.ts": 2472,
-	"packages/server/src/test/__tests__/migrate.test.ts": 4676,
-	"packages/server/src/test/__tests__/github.test.ts": 655,
-	"packages/server/src/test/__tests__/company-types.test.ts": 2984,
-	"packages/server/src/test/__tests__/issues.test.ts": 2923,
-	"packages/server/src/test/__tests__/project-docs.test.ts": 2960,
-	"packages/server/src/test/__tests__/db.test.ts": 3294,
-	"packages/server/src/test/__tests__/agent-api.test.ts": 3051,
-	"packages/server/src/test/__tests__/repos.test.ts": 3359,
-	"packages/server/src/test/__tests__/wakeup.test.ts": 2806,
-	"packages/server/src/test/__tests__/agent-types.test.ts": 2867,
-	"packages/server/src/test/__tests__/ssh-keys.test.ts": 2240,
-	"packages/server/src/test/__tests__/projects.test.ts": 2941,
-	"packages/server/src/test/__tests__/execution-locks.test.ts": 3110,
-	"packages/server/src/test/__tests__/secrets.test.ts": 2870,
-	"packages/server/src/test/__tests__/coach.test.ts": 3003,
-	"packages/server/src/test/__tests__/preview.test.ts": 2779,
-	"packages/server/src/test/__tests__/connections.test.ts": 3149,
-	"packages/server/src/test/__tests__/master-key.test.ts": 6581,
-	"packages/server/src/test/__tests__/ws.test.ts": 1166,
-	"packages/server/src/test/__tests__/cli.test.ts": 1809,
-	"packages/server/src/test/__tests__/git.test.ts": 2313,
-	"packages/server/src/test/__tests__/approvals.test.ts": 3490,
-	"packages/server/src/test/__tests__/startup.test.ts": 9134,
-	"packages/server/src/test/__tests__/api-keys.test.ts": 2585,
-	"packages/server/src/test/__tests__/audit-log.test.ts": 3068,
-	"packages/server/src/test/__tests__/template-resolver.test.ts": 3203,
-	"packages/connect/src/test/__tests__/config.test.ts": 997,
-	"packages/connect/src/test/__tests__/state.test.ts": 1062,
-	"packages/connect/src/test/__tests__/oauth.test.ts": 1097,
-	"packages/server/src/test/__tests__/companies.test.ts": 3340,
-	"packages/connect/src/test/__tests__/health.test.ts": 1061,
-	"packages/server/src/test/__tests__/state.test.ts": 4728,
-	"packages/server/src/test/__tests__/health.test.ts": 2887,
-	"packages/server/src/test/__tests__/costs.test.ts": 1994
+	"packages/server/src/test/__tests__/cli.test.ts": 5964,
+	"packages/connect/src/test/__tests__/health.test.ts": 6322,
+	"packages/server/src/test/__tests__/skills.test.ts": 20141,
+	"packages/server/src/test/__tests__/live-chat.test.ts": 20222,
+	"packages/server/src/test/__tests__/comments.test.ts": 20769,
+	"packages/server/src/test/__tests__/preview.test.ts": 20827,
+	"packages/server/src/test/__tests__/preferences.test.ts": 21417,
+	"packages/server/src/test/__tests__/agent-api.test.ts": 21637,
+	"packages/server/src/test/__tests__/execution-locks.test.ts": 17260,
+	"packages/server/src/test/__tests__/container-sync.test.ts": 17549,
+	"packages/server/src/test/__tests__/ws.test.ts": 2932,
+	"packages/server/src/test/__tests__/job-manager.test.ts": 4469,
+	"packages/server/src/test/__tests__/connections.test.ts": 31649,
+	"packages/server/src/test/__tests__/approvals.test.ts": 11092,
+	"packages/server/src/test/__tests__/repos.test.ts": 31889,
+	"packages/server/src/test/__tests__/projects.test.ts": 11769,
+	"packages/server/src/test/__tests__/audit-log.test.ts": 11208,
+	"packages/server/src/test/__tests__/template-resolver.test.ts": 11067,
+	"packages/server/src/test/__tests__/kb-docs.test.ts": 10889,
+	"packages/server/src/test/__tests__/migrate.test.ts": 16522,
+	"packages/server/src/test/__tests__/costs.test.ts": 13251,
+	"packages/server/src/test/__tests__/git.test.ts": 4862,
+	"packages/server/src/test/__tests__/issues.test.ts": 14604,
+	"packages/server/src/test/__tests__/company-types.test.ts": 12478,
+	"packages/server/src/test/__tests__/api-keys.test.ts": 12862,
+	"packages/server/src/test/__tests__/wakeup.test.ts": 13072,
+	"packages/server/src/test/__tests__/coach.test.ts": 14136,
+	"packages/server/src/test/__tests__/project-docs.test.ts": 13850,
+	"packages/server/src/test/__tests__/agent-types.test.ts": 14463,
+	"packages/connect/src/test/__tests__/state.test.ts": 3598,
+	"packages/connect/src/test/__tests__/config.test.ts": 4167,
+	"packages/server/src/test/__tests__/containers.test.ts": 4543,
+	"packages/server/src/test/__tests__/secrets.test.ts": 14097,
+	"packages/server/src/test/__tests__/ssh-keys.test.ts": 8650,
+	"packages/server/src/test/__tests__/db.test.ts": 13521,
+	"packages/server/src/test/__tests__/agents.test.ts": 12912,
+	"packages/server/src/test/__tests__/github.test.ts": 3420,
+	"packages/server/src/test/__tests__/state.test.ts": 25714,
+	"packages/connect/src/test/__tests__/oauth.test.ts": 3082,
+	"packages/server/src/test/__tests__/companies.test.ts": 12546,
+	"packages/server/src/test/__tests__/mcp.test.ts": 11769,
+	"packages/server/src/test/__tests__/health.test.ts": 7292,
+	"packages/server/src/test/__tests__/master-key.test.ts": 18068,
+	"packages/server/src/test/__tests__/startup.test.ts": 35972
 }
